+++ docs/docs/sandpack/plate-versions.ts
@@ -1,8 +1,9 @@
-export const plateVersion = '20.3.0';
+export const plateVersion = '20.5.0';
 export const testUtilsVersion = '13.4.0';
-export const excalidrawVersion = '20.0.0';
-export const juiceVersion = '20.0.0';
-export const selectionVersion = '20.0.0';
-export const cloudVersion = '20.0.0';
-export const cloudUiVersion = '20.0.0';
-export const dndVersion = '20.0.0';
+export const excalidrawVersion = '20.4.0';
+export const juiceVersion = '20.4.0';
+export const selectionVersion = '20.4.0';
+export const cloudVersion = '20.4.0';
+export const cloudUiVersion = '20.4.0';
+export const dndVersion = '20.4.0';
+export const dndUiVersion = '20.4.0';