+++ packages/floating/src/libs/radix-ui.ts
@@ -1,0 +1,1 @@
+export * as DropdownMenu from '@radix-ui/react-dropdown-menu';