+++ packages/ui/button/src/Button/PlateButton.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
+import React, { forwardRef } from 'react';
 import { Button } from '@udecode/plate-button';
 import { HTMLPropsAs } from '@udecode/plate-common';
-import { css } from 'styled-components';
+import { css, CSSProp } from 'styled-components';
 import tw from 'twin.macro';
 
 export interface PlateButtonProps extends HTMLPropsAs<'button'> {
@@ -11,10 +11,10 @@ export interface PlateButtonProps extends HTMLPropsAs<'button'> {
 }
 
 export const plateButtonCss = [
-  tw`relative inline-flex justify-center items-center text-center max-w-full p-0`,
-  tw`border-0 font-medium cursor-pointer`,
+  tw`relative inline-flex justify-center items-center text-center max-w-full p-0 box-border space-x-2`,
+  tw`border-0 font-medium cursor-pointer focus:!outline-none`,
   tw`bg-white hover:bg-gray-100 active:bg-gray-200`,
-  tw`px-2.5 py-1`,
+  tw`px-2.5 py-0 min-w-[28px] min-h-[28px]`,
   css`
     font-family: inherit;
     font-size: 14px;
@@ -29,7 +29,17 @@ export const plateButtonCss = [
     :visited {
       color: inherit;
     }
+
+    svg {
+      width: 16px;
+      height: 16px;
+    }
   `,
+];
+
+export const cssMenuItemButton: CSSProp = [
+  plateButtonCss,
+  tw`w-full justify-start`,
 ];
 
 export const primaryButtonCss = [
@@ -37,10 +47,8 @@ export const primaryButtonCss = [
   tw`bg-blue-500 text-white hover:bg-blue-600 active:bg-blue-700 active:text-white`,
 ];
 
-export const PlateButton = ({
-  size,
-  px,
-  py,
-  css: _css,
-  ...props
-}: PlateButtonProps) => <Button css={plateButtonCss} {...props} />;
+export const PlateButton = forwardRef<HTMLButtonElement, PlateButtonProps>(
+  ({ size, px, py, css: _css, ...props }, ref) => (
+    <Button type="button" ref={ref} css={plateButtonCss} {...props} />
+  )
+);