+++ packages/ui/nodes/table/src/PlateTableElement.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
-import { getPluginOptions, usePlateEditorRef } from '@udecode/plate-common';
 import {
-  ELEMENT_TABLE,
   TableElement,
   TableElementRootProps,
-  TablePlugin,
   useTableElementState,
 } from '@udecode/plate-table';
 import { css } from 'styled-components';
@@ -14,41 +11,45 @@ import { PlateTablePopover } from './PlateTablePopover';
 export const PlateTableElement = (props: TableElementRootProps) => {
   const { as, children, ...rootProps } = props;
 
-  const editor = usePlateEditorRef();
-  const { minColumnWidth: minWidth } = getPluginOptions<TablePlugin>(
-    editor,
-    ELEMENT_TABLE
-  );
-  const { colSizes, isSelectingCell } = useTableElementState();
+  const {
+    colSizes,
+    isSelectingCell,
+    minColumnWidth,
+    marginLeft,
+  } = useTableElementState();
 
   return (
-    <TableElement.Root
-      css={[
-        tw`table table-fixed w-full h-px my-4 ml-px mr-0 border-collapse`,
-        isSelectingCell &&
-          css`
-            *::selection {
-              background: none;
-            }
-          `,
-      ]}
-      {...rootProps}
-    >
-      <TableElement.ColGroup>
-        {colSizes.map((width, index) => (
-          <TableElement.Col
-            key={index}
-            style={{
-              minWidth,
-              width: width || undefined,
-            }}
-          />
-        ))}
-      </TableElement.ColGroup>
+    <PlateTablePopover>
+      <TableElement.Wrapper style={{ paddingLeft: marginLeft }}>
+        <TableElement.Root
+          css={[
+            tw`table table-fixed w-full h-px my-4 ml-px mr-0 border-collapse`,
+            isSelectingCell &&
+              css`
+                *::selection {
+                  background: none;
+                }
+              `,
+          ]}
+          {...rootProps}
+        >
+          <TableElement.ColGroup>
+            {colSizes.map((width, index) => (
+              <TableElement.Col
+                key={index}
+                style={{
+                  minWidth: minColumnWidth,
+                  width: width || undefined,
+                }}
+              />
+            ))}
+          </TableElement.ColGroup>
 
-      <PlateTablePopover>
-        <TableElement.TBody css={tw`min-w-full`}>{children}</TableElement.TBody>
-      </PlateTablePopover>
-    </TableElement.Root>
+          <TableElement.TBody css={tw`min-w-full`}>
+            {children}
+          </TableElement.TBody>
+        </TableElement.Root>
+      </TableElement.Wrapper>
+    </PlateTablePopover>
   );
 };