+++ packages/ui/plate/src/index.tsx
@@ -5,6 +5,7 @@
 export * from '@udecode/plate-styled-components';
 export * from '@udecode/plate-ui-alignment';
 export * from '@udecode/plate-ui-block-quote';
+export * from '@udecode/plate-ui-button';
 export * from '@udecode/plate-ui-code-block';
 export * from '@udecode/plate-ui-combobox';
 export * from '@udecode/plate-ui-comments';
@@ -21,6 +22,5 @@ export * from '@udecode/plate-ui-mention';
 export * from '@udecode/plate-ui-placeholder';
 export * from '@udecode/plate-ui-table';
 export * from '@udecode/plate-ui-toolbar';
-export * from '@udecode/plate-ui-button';
 export * from './types';
 export * from './utils';