+++ packages/cloud/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-cloud
 
+## 20.4.0
+
+## 20.3.2
+
 ## 20.0.0
 
 ### Minor Changes