+++ packages/editor/autoformat/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @udecode/plate-autoformat
 
+## 20.4.0
+
+## 20.3.2
+
+### Patch Changes
+
+- [#2285](https://github.com/udecode/plate/pull/2285) by [@12joan](https://github.com/12joan) – Ignore `defaultPrevented` keydown events
+
 ## 20.0.0
 
 ## 19.7.0