+++ packages/ui/cursor/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-ui-cursor
 
+## 20.4.0
+
+## 20.3.2
+
 ## 20.0.1
 
 ### Patch Changes