+++ packages/ui/nodes/horizontal-rule/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @udecode/plate-horizontal-rule-ui
 
+## 20.4.0
+
+## 20.3.2
+
 ## 20.0.0
 
 ## 19.7.0